+++ src/main/java/edu/wpi/rail/jrosbridge/core/Ros.java
@@ -34,7 +34,7 @@ import edu.wpi.rail.jrosbridge.core.callback.TopicCallback;
  * edu.wpi.rail.jrosbridge.JRosbridge.Topic, are used.
  * 
  * @author Russell Toris - rctoris@wpi.edu
- * @version Feb. 16, 2014
+ * @version Feb. 18, 2014
  */
 @ClientEndpoint
 public class Ros {
@@ -51,6 +51,7 @@ public class Ros {
 
 	private String hostname;
 	private int port;
+	private JRosbridge.WebSocketType protocol;
 
 	// active session (stored upon connection)
 	private Session session;
@@ -77,7 +78,8 @@ public class Ros {
 
 	/**
 	 * Create a connection to ROS with the given hostname and default port. A
-	 * call to connect must be made to establish a connection.
+	 * call to connect must be made to establish a connection. By default,
+	 * WebSockets is used (as opposed to WSS).
 	 * 
 	 * @param hostname
 	 *            The hostname to connect to.
@@ -88,7 +90,8 @@ public class Ros {
 
 	/**
 	 * Create a connection to ROS with the given hostname and port. A call to
-	 * connect must be made to establish a connection.
+	 * connect must be made to establish a connection. By default, WebSockets is
+	 * used (as opposed to WSS).
 	 * 
 	 * @param hostname
 	 *            The hostname to connect to.
@@ -96,8 +99,24 @@ public class Ros {
 	 *            The port to connect to.
 	 */
 	public Ros(String hostname, int port) {
+		this(hostname, port, JRosbridge.WebSocketType.ws);
+	}
+
+	/**
+	 * Create a connection to ROS with the given hostname and port. A call to
+	 * connect must be made to establish a connection.
+	 * 
+	 * @param hostname
+	 *            The hostname to connect to.
+	 * @param port
+	 *            The port to connect to.
+	 * @param protocol
+	 *            The WebSocket protocol to use.
+	 */
+	public Ros(String hostname, int port, JRosbridge.WebSocketType protocol) {
 		this.hostname = hostname;
 		this.port = port;
+		this.protocol = protocol;
 		this.session = null;
 		this.idCounter = 0;
 		this.topicCallbacks = new HashMap<String, ArrayList<TopicCallback>>();
@@ -121,6 +140,25 @@ public class Ros {
 	 */
 	public int getPort() {
 		return this.port;
+	}
+
+	/**
+	 * Get the type of WebSocket protocol being used.
+	 * 
+	 * @return The type of WebSocket protocol being used.
+	 */
+	public JRosbridge.WebSocketType getProtocol() {
+		return this.protocol;
+	}
+
+	/**
+	 * Get the full URL this client is connecting to.
+	 * 
+	 * @return
+	 */
+	public String getURL() {
+		return this.protocol.toString() + "://" + this.hostname + ":"
+				+ this.port;
 	}
 
 	/**
@@ -153,7 +191,7 @@ public class Ros {
 	public boolean connect() {
 		try {
 			// create a WebSocket connection here
-			URI uri = new URI("ws://" + hostname + ":" + port);
+			URI uri = new URI(this.getURL());
 			ContainerProvider.getWebSocketContainer()
 					.connectToServer(this, uri);
 			return true;
@@ -192,7 +230,7 @@ public class Ros {
 	 * @return If there is a connection to rosbridge.
 	 */
 	public boolean isConnected() {
-		return session != null && session.isOpen();
+		return this.session != null && this.session.isOpen();
 	}
 
 	/**
@@ -362,6 +400,38 @@ public class Ros {
 	}
 
 	/**
+	 * Sends an authorization request to the server.
+	 * 
+	 * @param mac
+	 *            The MAC (hash) string given by the trusted source.
+	 * @param client
+	 *            The IP of the client.
+	 * @param dest
+	 *            The IP of the destination.
+	 * @param rand
+	 *            The random string given by the trusted source.
+	 * @param t
+	 *            The time of the authorization request.
+	 * @param level
+	 *            The user level as a string given by the client.
+	 * @param end
+	 *            The end time of the client's session.
+	 */
+	public void authenticate(String mac, String client, String dest,
+			String rand, int t, String level, int end) {
+		// build and send the rosbridge call
+		JsonObject call = Json.createObjectBuilder()
+				.add(JRosbridge.FIELD_OP, JRosbridge.OP_CODE_AUTHENTICATE)
+				.add(JRosbridge.FIELD_MAC, mac)
+				.add(JRosbridge.FIELD_CLIENT, client)
+				.add(JRosbridge.FIELD_DESTINATION, dest)
+				.add(JRosbridge.FIELD_RAND, rand).add(JRosbridge.FIELD_TIME, t)
+				.add(JRosbridge.FIELD_LEVEL, level)
+				.add(JRosbridge.FIELD_END_TIME, end).build();
+		this.send(call);
+	}
+
+	/**
 	 * Register a callback for a given topic.
 	 * 
 	 * @param topic
@@ -419,7 +489,9 @@ public class Ros {
 	}
 
 	public static void main(String[] args) throws InterruptedException {
+
 		Ros ros = new Ros("rct-desktop.cs.wpi.edu");
+		System.out.println(ros.getURL());
 		ros.connect();
 
 		Topic echo = new Topic(ros, "/echo", "std_msgs/String");