+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 DEVEL - **0.0.3**
  * Support for parameters added [(rctoris)](https://github.com/rctoris/)
- * `edu.wpi.rail.jrosbridge.math package` added [(rctoris)](https://github.com/rctoris/)
+ * `edu.wpi.rail.jrosbridge.math` package added [(rctoris)](https://github.com/rctoris/)
  * Callbacks added to their own `edu.wpi.rail.jrosbridge.core.callback` package [(rctoris)](https://github.com/rctoris/)
  * png decompression support added [(rctoris)](https://github.com/rctoris/)
+ * WSS client connections now supported [(rctoris)](https://github.com/rctoris/)
+ * `rosauth` authentication added [(rctoris)](https://github.com/rctoris/)
 
 2013-02-16 - **0.0.2**
  * Support for calling services added [(rctoris)](https://github.com/rctoris/)